Some comments of errors
Two issues with the code, while the claw is in motion (aka was selected and is moving to the closed/open position) the user has the ability to continuously submit the move which doubles the speed every time. Second issue, whenever you reset the robot, the claw resets to the wrong position (should reset to open but will sometimes reset to a half-open state).